Add support for filters that have exclusion rules
A filter that is an array of [regex/string, regex, regex] will use the first element as a normal filter, and then exclude the match if the text before the matching phrase is matched by the second regex or if the text after the matching phrase is matched by the third regex.
The second and third regex are optional, but at least one must be provided ([regex/string, regex] is valid and [regex/string, null, regex] is valid)

A filter that is an array of [regex/string, function] will use the first element as a normal filter, and then exclude the match if the function returns true.
The function will be passed the following parameters: phrase, beforePhrase, afterPhrase, startIndex, fullText

Contributes towards #25